pyproject.toml
[tool.poetry]
name = "cat-dog-telegram"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "cat_dog_telegram"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
tensorflow = "^2.11.0"
keras = "^2.11.0"
numpy = "^1.24.1"
pillow = "^9.4.0"
aiokafka = "^0.8.0"
aiogram = "^2.24"
python-dotenv = "^0.21.0"
scipy = "^1.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
